--- gateway/src/apicast/policy/mtls/init.lua
@@ -1,1 +1,0 @@
-return require('mtls')
