+++ gateway/src/apicast/policy/oauth_mtls/init.lua
@@ -1,0 +1,1 @@
+return require('oauth_mtls')
